# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/login": {
            "post": {
                "description": "账号登陆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "账号登陆",
                "parameters": [
                    {
                        "description": "登陆信息",
                        "name": "userInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.LoginInfos"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login/code": {
            "get": {
                "description": "获取错误代码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "获取错误代码",
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login/info": {
            "get": {
                "description": "获取登陆信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "获取登陆信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/login/register": {
            "put": {
                "description": "账号申请",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "账号申请",
                "parameters": [
                    {
                        "description": "申请的用户信息",
                        "name": "userInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Registerinfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Result\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Result\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login/register/name": {
            "get": {
                "description": "检测账号是否存在",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "检测账号是否存在",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.LoginInfos": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "验证码",
                    "type": "string"
                },
                "name": {
                    "description": "登陆名",
                    "type": "string"
                },
                "password": {
                    "description": "密码 MD5加密(小写)",
                    "type": "string"
                }
            }
        },
...
        "api.Registerinfo": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "验证码",
                    "type": "string"
                },
                "codeJWT": {
                    "description": "验证码标记",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "mobile": {
                    "description": "手机号码",
                    "type": "string"
                },
                "name": {
                    "description": "账号",
                    "type": "string"
                },
                "password": {
                    "description": "密码 MD5加密(小写)",
                    "type": "string"
                }
            }
        },
...
    }
}